refactor : reflecting feedback
