+++ src/main/java/flab/payment_system/config/QueryDslConfig.java
@@ -7,7 +7,7 @@ import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 
 @Configuration
-public class QueryDslConfiguration {
+public class QueryDslConfig {
 
 	@PersistenceContext
 	private EntityManager entityManager;
