--- src/main/java/flab/payment_system/config/WebMvcConfiguration.java
@@ -1,22 +1,0 @@
-package flab.payment_system.config;
-
-
-import flab.payment_system.core.interceptor.LoggingInterceptor;
-import lombok.RequiredArgsConstructor;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
-import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-
-@Configuration
-@RequiredArgsConstructor
-public class WebMvcConfiguration implements WebMvcConfigurer {
-
-	private final LoggingInterceptor loggingInterceptor;
-
-	@Override
-	public void addInterceptors(InterceptorRegistry registry) {
-		registry.addInterceptor(loggingInterceptor)
-			.order(1)
-			.addPathPatterns("/**");
-	}
-}
