+++ src/main/java/flab/payment_system/domain/order/dto/OrderDetailDto.java
@@ -3,3 +3,4 @@ package flab.payment_system.domain.order.dto;
 public record OrderDetailDto(String tid) {
 
 }
+
